AfroBench/prompt_with_API/requirements.txt
aiolimiter~=1.1.0
aiosignal~=1.3.1
annotated-types~=0.7.0
genai~=2.1.0
google-generativeai~=0.8.3
openai~=0.27.10
rouge-score~=0.1.2
sacrebleu~=2.4.3
together~=1.3.14
tqdm~=4.66.5
tqdm-multiprocess~=0.0.11
PyYAML~=6.0.2
pandas~=2.2.3
numpy~=2.2.4
datasets~=3.4.1
aiohttp~=3.11.14
scikit-learn~=1.6.1
evaluate~=0.4.3
sentencepiece~=0.2.0
torch~=2.6.0
bert-score~=0.3.13